https://www.kinfolkedu.com/index.html
Inspired by the Montessori method and crafted into an innovative and holistic early childhood education program that provides a strong foundation aimed at continuing all-round development progressive through primary and secondary school levels. Our motto is “Follow the Child” and the entire environment is designed in keeping with the ethos.